Crate odbc_sys

ODBC types those representation is compatible with the ODBC C API.

This layer has not been created using automatic code generation. It is incomplete, i.e. it does not contain every symbol or constant defined in the ODBC C headers. Symbols which are deprecated since ODBC 3 have been left out intentionally. While some extra type safety has been added by grouping some of C's #define constants into enum-types it mostly offers the same power (all) and safety guarantess(none) as the wrapped C-API. ODBC 4.0 is still under development by Microsoft, so these symbols are deactivated by default in the cargo.toml

Reexports

pub use self::SQLRETURN::*;

pub use self::InfoType::*;

pub use self::FetchOrientation::*;

pub use EnvironmentAttribute::*;

pub use OdbcVersion::*;

pub use self::SqlCDataType::*;

pub use SQL_C_ULONG as SQL_C_BOOKMARK;

pub use InputOutput::*;

pub use Nullable::*;

pub use FreeStmtOption::*;

pub use self::SqlDataType::*;

pub use self::HandleType::*;

pub use self::SqlDriverConnectOption::*;

pub use self::SqlStatementAttribute::*;

pub use self::SqlConnectionAttribute::*;

pub use self::SqlCompletionType::*;

Enums

Dbc

Env

EnvironmentAttribute
Governs behaviour of EnvironmentAttribute

FetchOrientation
Codes used for FetchOrientation in SQLFetchScroll, SQLDataSources and in SQLDrivers

FreeStmtOption
SQL Free Statement options

HandleType
Represented in C headers as SQLSMALLINT

InfoType
Information requested by SQLGetInfo

InputOutput
Used by SQLBindParameter.

Nullable
Used by SQLDescribeCol.

Obj

OdbcVersion
ODBC verions

SQLRETURN
Indicates the overall success or failure of the function

SqlCDataType
The C data type is specified in the SQLBindCol and SQLGetData functions with the TargetType argument and in the SQLBindParameter function with the ValueType argument.

SqlCompletionType
Completion types for SQLEndTrans

SqlConnectionAttribute
Connection attributes for SQLSetConnectAttr

SqlDataType
SQL Data Types

SqlDriverConnectOption
Options for SQLDriverConnect

SqlStatementAttribute
Statement attributes for SQLSetStmtAttr

Stmt

Constants

SQL_MAX_MESSAGE_LENGTH
Maximum message length

SQL_NO_TOTAL

SQL_NTS

SQL_NTSL

SQL_NULL_DATA

SQL_SQLSTATE_SIZE

Functions

SQLAllocHandle⚠
Allocates an environment, connection, statement, or descriptor handle.

SQLBindCol⚠
Binds application data bufferst to columns in the result set.

SQLBindParameter⚠
Binds a buffer to a parameter marker in an SQL statement

SQLCloseCursor⚠
Closes a cursor that has been opened on a statement and discards pending results.

SQLConnect⚠
SQLConnect establishes connections to a driver and a data source. The connection handle references storage of all information about the connection to the data source, including status, transaction state, and error information.

SQLDataSources⚠
Returns information about a data source. This function is implemented only by the Driver Manager.

SQLDescribeCol⚠
Returns the result descriptor for one column in the result set — column name, type, column size, decimal digits, and nullability.

SQLDisconnect⚠
Closes the connection associated with a specific connection handle.

SQLDriverConnect⚠
An alternative to SQLConnect. It supports data sources that require more connection information than the three arguments in SQLConnect, dialog boxes to prompt the user for all connection information, and data sources that are not defined in the system information

SQLDrivers⚠
Lists driver descriptions and driver attribute keywords. This function is implemented only by the Driver Manager.

SQLEndTran⚠
Requests a commit or rollback operation for all active operations on all statements associated with a handle.

SQLExecDirect⚠
Executes a preparable statement, using the current values of the parameter marker variables if any parameters exist in the statement. This is the fastest way to submit an SQL statement for one-time execution

SQLExecute⚠
Executes a prepared statement, using the current values of the parameter marker variables if any paramater markers exis in the statement.

SQLFetch⚠
SQLFetch fetches the next rowset of data from the result set and returns data for all bound columns.

SQLFreeHandle⚠
Frees resources associated with a specific environment, connection, statement, or descriptor handle.

SQLFreeStmt⚠
Stops processing associated with a specific statement, closes any open cursors associated with the statement, discards pending results, or, optionally, frees all resources associated with the statement handle.

SQLGetData⚠

SQLGetDiagRec⚠
Return the current values of multiple fields of a diagnostic record that contains eror, warning, and status information.

SQLGetInfo⚠
Returns general information about the driver and data source associated with a connection

SQLNumResultCols⚠
Returns the number of columns in a result set

SQLPrepare⚠
Compiles the statement and generates an access plan.

SQLSetConnectAttr⚠
Sets attributes that govern aspects of connections.

SQLSetEnvAttr⚠
Sets attributes that govern aspects of environments

SQLSetStmtAttr⚠
Sets attributes related to a statement.

SQLTables⚠
Returns the list of table, catalog, or schema names, and table types, stored in a specific data source. The driver returns the information as a result set

Type Definitions

SQLCHAR

SQLHANDLE

SQLHDBC
The connection handle references storage of all information about the connection to the data source, including status, transaction state, and error information.

SQLHENV

SQLHSTMT

SQLHWND

SQLINTEGER

SQLLEN

SQLPOINTER

SQLSMALLINT

SQLUINTEGER

SQLULEN

SQLUSMALLINT