Enum odbc_sys::SQLRETURN

#[repr(i16)]
#[must_use]

pub enum SQLRETURN {
    SQL_INVALID_HANDLE,
    SQL_ERROR,
    SQL_SUCCESS,
    SQL_SUCCESS_WITH_INFO,
    SQL_STILL_EXECUTING,
    SQL_NEED_DATA,
    SQL_NO_DATA,
    SQL_PARAM_DATA_AVAILABLE,
}

Indicates the overall success or failure of the function

Each function in ODBC returns a code, known as its return code, which indicates the overall success or failure of the function. Program logic is generally based on return codes. See ODBC reference

Variants
SQL_INVALID_HANDLE

Function failed due to an invalid environment, connection, statement, or descriptor handle

This indicates a programming error. No additional information is available from SQLGetDiagRec or SQLGetDiagField. This code is returned only when the handle is a null pointer or is the wrong type, such as when a statement handle is passed for an argument a connection handle.

SQL_ERROR

Function failed

The application calls SQLGetDiagRec or SQLGetDiagField to retrieve additional information. The contents of any output arguments to the function are undefined.

SQL_SUCCESS

Function completed successfully

The application calls SQLGetDiagField to retrieve additional information from the header record.

SQL_SUCCESS_WITH_INFO

Function completed successfully, possibly with a nonfatal error (warning)

The application calls SQLGetDiagRec or SQLGetDiagField to retrieve additional information.

SQL_STILL_EXECUTING

A function that was started asynchronously is still executing

The application SQLGetDiagRec or SQLGetDiagField to retrieve additional information if any.

SQL_NEED_DATA

More data is needed

,such as when a parameter data is sent at execution time or additional connection information is required. The application calls SQLGetDiagRec or SQLGetDiagField to retrieve additional information, if any.

SQL_NO_DATA

No more data was available

The application calls SQLGetDiagRec or SQLGetDiagField to retrieve additional information. One or more driver-defined status records in class 02xxx may be returned.

SQL_PARAM_DATA_AVAILABLE