Function odbc_sys::SQLDriverConnect

pub unsafe extern "C" fn SQLDriverConnect(
    connection_handle: SQLHDBC, 
    window_handle: SQLHWND, 
    in_connection_string: *const SQLCHAR, 
    string_length_1: SQLSMALLINT, 
    out_connection_string: *mut SQLCHAR, 
    buffer_length: SQLSMALLINT, 
    string_length_2: *mut SQLSMALLINT, 
    DriverCompletion: SqlDriverConnectOption
) -> SQLRETURN

An alternative to SQLConnect. It supports data sources that require more connection information than the three arguments in SQLConnect, dialog boxes to prompt the user for all connection information, and data sources that are not defined in the system information

Returns

SQL_SUCCESS, SQL_SUCCESS_WITH_INFO, SQL_ERROR, SQL_INVALID_HANDLE, SQL_NO_DATA, or SQL_STILL_EXECUTING