Crate odbc_sys

ODBC types those representation is compatible with the ODBC C API.

This layer has not been created using automatic code generation. It is incomplete, i.e. it does not contain every symbol or constant defined in the ODBC C headers. Symbols which are deprecated since ODBC 3 have been left out intentionally. While some extra type safety has been added by grouping some of C's #define constants into enum-types it mostly offers the same power (all) and safety guarantess(none) as the wrapped C-API. ODBC 4.0 is still under development by Microsoft, so these symbols are deactivated by default in the cargo.toml

Re-exports

pub use EnvironmentAttribute::*;

pub use OdbcVersion::*;

pub use self::SqlCDataType::*;

pub use SQL_C_ULONG as SQL_C_BOOKMARK;

pub use self::FetchOrientation::*;

pub use self::InfoType::*;

pub use InputOutput::*;

pub use Nullable::*;

pub use self::SqlBulkOperation::*;

pub use self::SQLRETURN::*;

pub use FreeStmtOption::*;

pub use self::SqlDataType::*;

pub use self::HandleType::*;

pub use self::SqlDriverConnectOption::*;

pub use self::SqlAttributeStringLength::*;

pub use self::SqlStatementAttribute::*;

pub use self::SqlAsyncConnectionBehavior::*;

pub use self::SqlConnectionAttribute::*;

pub use self::SqlCompletionType::*;

Structs

SQLGUID

SQL_DATE_STRUCT

SQL_DAY_SECOND_STRUCT

SQL_INTERVAL_STRUCT

SQL_SS_TIME2_STRUCT

SQL_SS_TIMESTAMPOFFSET_STRUCT

SQL_TIMESTAMP_STRUCT

SQL_TIME_STRUCT

SQL_YEAR_MONTH_STRUCT

Enums

Dbc

Desc

Env

EnvironmentAttribute
Governs behaviour of EnvironmentAttribute

FetchOrientation
Codes used for FetchOrientation in SQLFetchScroll, SQLDataSources and in SQLDrivers

FreeStmtOption
SQL Free Statement options

HandleType
Represented in C headers as SQLSMALLINT

InfoType
Information requested by SQLGetInfo

InputOutput
Used by SQLBindParameter.

Nullable
Used by SQLDescribeCol.

Obj

OdbcVersion
ODBC verions

SQLINTERVAL

SQLRETURN
Indicates the overall success or failure of the function

SqlAsyncConnectionBehavior

SqlAttributeStringLength

SqlBulkOperation
Supported SQLBulkOperations operations

SqlCDataType
The C data type is specified in the SQLBindCol and SQLGetData functions with the TargetType argument and in the SQLBindParameter function with the ValueType argument.

SqlCompletionType
Completion types for SQLEndTrans

SqlConnectionAttribute
Connection attributes for SQLSetConnectAttr

SqlDataType
SQL Data Types

SqlDriverConnectOption
Options for SQLDriverConnect

SqlDynamicDiagnosticIdentifier

SqlHeaderDiagnosticIdentifier
DiagIdentifier for SQLGetDiagField

SqlStatementAttribute
Statement attributes for SQLSetStmtAttr

Stmt

Constants

SQL_C_TYPES_EXTENDED
Extended C Types range 4000 and above. Range of -100 thru 200 is reserved by Driver Manager.

SQL_MAX_MESSAGE_LENGTH
Maximum message length

SQL_NO_TOTAL

SQL_NTS

SQL_NTSL

SQL_NULL_DATA

SQL_SQLSTATE_SIZE

SQL_SQLSTATE_SIZEW

SQL_SS_LENGTH_UNLIMITED

Functions

SQLAllocHandle⚠
Allocates an environment, connection, statement, or descriptor handle.

SQLBindCol⚠
Binds application data bufferst to columns in the result set.

SQLBindParameter⚠
Binds a buffer to a parameter marker in an SQL statement

SQLBrowseConnectW⚠
SQLBrowseConnect supports an iterative method of discovering and enumerating the attributes and attribute values required to connect to a data source. Each call to SQLBrowseConnect returns successive levels of attributes and attribute values.

SQLBulkOperations⚠
Performs bulk insertions and bulk bookmark operations, including update, delete, and fetch by bookmark.

SQLCancel⚠
Cancels the processing on a statement.

SQLCancelHandle⚠
Cancels the processing on a connection or statement.

SQLCloseCursor⚠
Closes a cursor that has been opened on a statement and discards pending results.

SQLColAttributeW⚠
Returns descriptor information for a column in a result set. Descriptor information is returned as a character string, a descriptor-dependent value, or an integer value.

SQLColumnPrivilegesW⚠
Returns a list of columns and associated privileges for the specified table. The driver returns the information as a result set on the specified StatementHandle.

SQLColumnsW⚠
Returns the list of column names in specified tables. The driver returns this information as a result set on the specified StatementHandle.

SQLCompleteAsync⚠
Can be used to determine when an asynchronous function is complete using either notification- or polling-based processing.

SQLConnect⚠
SQLConnect establishes connections to a driver and a data source. The connection handle references storage of all information about the connection to the data source, including status, transaction state, and error information.

SQLConnectW⚠
SQLConnect establishes connections to a driver and a data source. The connection handle references storage of all information about the connection to the data source, including status, transaction state, and error information.

SQLCopyDesc⚠
Copies descriptor information from one descriptor handle to another.

SQLDataSources⚠
Returns information about a data source. This function is implemented only by the Driver Manager.

SQLDataSourcesW⚠
Returns information about a data source. This function is implemented only by the Driver Manager.

SQLDescribeCol⚠
Returns the result descriptor for one column in the result set — column name, type, column size, decimal digits, and nullability.

SQLDescribeColW⚠
Returns the result descriptor for one column in the result set — column name, type, column size, decimal digits, and nullability.

SQLDescribeParam⚠
Returns the description of a parameter marker associated with a prepared SQL statement. This information is also available in the fields of the IPD.

SQLDisconnect⚠
Closes the connection associated with a specific connection handle.

SQLDriverConnect⚠
An alternative to SQLConnect. It supports data sources that require more connection information than the three arguments in SQLConnect, dialog boxes to prompt the user for all connection information, and data sources that are not defined in the system information

SQLDriverConnectW⚠
An alternative to SQLConnect. It supports data sources that require more connection information than the three arguments in SQLConnect, dialog boxes to prompt the user for all connection information, and data sources that are not defined in the system information

SQLDrivers⚠
Lists driver descriptions and driver attribute keywords. This function is implemented only by the Driver Manager.

SQLDriversW⚠
Lists driver descriptions and driver attribute keywords. This function is implemented only by the Driver Manager.

SQLEndTran⚠
Requests a commit or rollback operation for all active operations on all statements associated with a handle.

SQLExecDirect⚠
Executes a preparable statement, using the current values of the parameter marker variables if any parameters exist in the statement. This is the fastest way to submit an SQL statement for one-time execution

SQLExecDirectW⚠
Executes a preparable statement, using the current values of the parameter marker variables if any parameters exist in the statement. This is the fastest way to submit an SQL statement for one-time execution

SQLExecute⚠
Executes a prepared statement, using the current values of the parameter marker variables if any paramater markers exis in the statement.

SQLFetch⚠
SQLFetch fetches the next rowset of data from the result set and returns data for all bound columns.

SQLFetchScroll⚠
Fetches the specified rowset of data from the result set and returns data for all bound columns. Rowsets can be specified at an absolute or relative position or by bookmark.

SQLForeignKeysW⚠
Can return:

SQLFreeHandle⚠
Frees resources associated with a specific environment, connection, statement, or descriptor handle.

SQLFreeStmt⚠
Stops processing associated with a specific statement, closes any open cursors associated with the statement, discards pending results, or, optionally, frees all resources associated with the statement handle.

SQLGetConnectAttrW⚠
Returns the current setting of a connection attribute.

SQLGetCursorNameW⚠
Returns the cursor name associated with a specified statement.

SQLGetData⚠

SQLGetDescFieldW⚠
Returns the current setting or value of a single field of a descriptor record.

SQLGetDescRecW⚠
Returns the current settings or values of multiple fields of a descriptor record. The fields returned describe the name, data type, and storage of column or parameter data.

SQLGetDiagFieldW⚠
Returns the current value of a field of a record of the diagnostic data structure (associated with a specified handle) that contains error, warning, and status information.

SQLGetDiagRec⚠
Return the current values of multiple fields of a diagnostic record that contains eror, warning, and status information.

SQLGetDiagRecW⚠
Return the current values of multiple fields of a diagnostic record that contains eror, warning, and status information.

SQLGetEnvAttr⚠
Gets attributes that govern aspects of environments

SQLGetInfo⚠
Returns general information about the driver and data source associated with a connection

SQLGetInfoW⚠
Returns general information about the driver and data source associated with a connection

SQLGetStmtAttrW⚠
Returns the current setting of a statement attribute.

SQLGetTypeInfo⚠
SQLGetTypeInfo returns information about data types supported by the data source. The driver returns the information in the form of an SQL result set. The data types are intended for use in Data Definition Language (DDL) statements.

SQLMoreResults⚠
Determines whether more results are available on a statement containing SELECT, UPDATE, INSERT, or DELETE statements and, if so, initializes processing for those results.

SQLNumResultCols⚠
Returns the number of columns in a result set

SQLPrepare⚠
Compiles the statement and generates an access plan.

SQLPrepareW⚠
Compiles the statement and generates an access plan.

SQLRowCount⚠
Returns the number of rows affected by an UPDATE, INSERT, or DELETE statement; an SQL_ADD, SQL_UPDATE_BY_BOOKMARK, or SQL_DELETE_BY_BOOKMARK operation in SQLBulkOperations; or an SQL_UPDATE or SQL_DELETE operation in SQLSetPos.

SQLSetConnectAttr⚠
Sets attributes that govern aspects of connections.

SQLSetConnectAttrW⚠
Sets attributes that govern aspects of connections.

SQLSetEnvAttr⚠
Sets attributes that govern aspects of environments

SQLSetStmtAttr⚠
Sets attributes related to a statement.

SQLSetStmtAttrW⚠
Sets attributes related to a statement.

SQLTables⚠
Returns the list of table, catalog, or schema names, and table types, stored in a specific data source. The driver returns the information as a result set

SQLTablesW⚠
Returns the list of table, catalog, or schema names, and table types, stored in a specific data source. The driver returns the information as a result set

Type Definitions

RETCODE

SQLCHAR

SQLHANDLE

SQLHDBC
The connection handle references storage of all information about the connection to the data source, including status, transaction state, and error information.

SQLHDESC

SQLHENV

SQLHSTMT

SQLHWND

SQLINTEGER

SQLLEN

SQLPOINTER

SQLSMALLINT

SQLUINTEGER

SQLULEN

SQLUSMALLINT

SQLWCHAR

Unions

SQL_INTERVAL_UNION